Tests for sorting add
